app/requirements.txt
asgiref==3.4.1
Django==3.2.13
django-phone-field==1.8.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
zipp==3.6.0
